Revert "adding support for azure metrics"
This reverts commit 7731ad38faaf5287b9317bd9141ae15875eefde4.
