--- eventHubFunction/logzioMetricsFunction/index.js
@@ -1,5 +1,0 @@
-const logsProcessor = require('../SharedCode/mainIndex');
-
-module.exports = function processEventHubMessages(context, eventHubMessages) {
-  logsProcessor(context, eventHubMessages, true);
-};
