+++ eventHubFunction/package.json
@@ -22,7 +22,6 @@
     "eslint-plugin-import": "^2.14.0"
   },
   "dependencies": {
-    "lodash": "^4.17.11",
     "logzio-nodejs": "^1.0.3"
   }
 }
