+++ eventHubFunctionTests/test-logs.js
