--- test/test-metrics.js
@@ -1,15 +1,0 @@
-const aksAgentPool = {
-  count: 1,
-  total: 0,
-  average: 0,
-  minimum: 0,
-  maximum: 0,
-  resourceId: '/SUBSCRIPTIONS/6EFCEE25/RESOURCEGROUPS/MC_LOGZIO_MYCLUSTER_EA/PROVIDERS/MICROSOFT.NETWORK/NETWORKINTERFACES/AKS-AGENTPOOL-2-NIC-0',
-  time: '2019-01-06T12:35:00.000Z',
-  metricName: 'BytesSentRate',
-  timeGrain: 'PT1M',
-};
-
-module.exports = {
-  aksAgentPool,
-};
